Split `Fill` and `Stroke` into their own modules
