+++ wgpu/src/widget/canvas/fill.rs
@@ -1,0 +1,12 @@
+use iced_native::Color;
+
+#[derive(Debug, Clone, Copy)]
+pub enum Fill {
+    Color(Color),
+}
+
+impl Default for Fill {
+    fn default() -> Fill {
+        Fill::Color(Color::BLACK)
+    }
+}
